backend/requirements.txt
# Core Framework
agno>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# AI/ML Dependencies
google-generativeai>=0.3.0
openai>=1.3.0

# Memory Systems
zep-python>=0.30.0
mem0ai
redis>=5.0.0

# Web Framework & API
websockets>=12.0
python-multipart>=0.0.6

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Environment & Configuration
python-dotenv>=1.0.0

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Development Tools
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Frontend Integration
jinja2>=3.1.0
aiofiles>=23.2.0

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 